+++ libs/react-components/src/community/components/card/card.stories.tsx
@@ -1,10 +1,10 @@
 import { Meta, StoryFn, StoryObj } from '@storybook/react';
 
 import { Col, Row } from '../../../tedi/components/grid';
+import Separator from '../../../tedi/components/separator/separator';
 import { VerticalSpacing } from '../../../tedi/components/vertical-spacing';
 import Collapse from '../collapse/collapse';
 import Icon from '../icon/icon';
-import Separator from '../separator/separator';
 import StretchContent from '../stretch-content/stretch-content';
 import { CardsExample } from '../stretch-content/stretch-content.stories';
 import Heading from '../typography/heading/heading';
@@ -333,7 +333,7 @@ const Timeline: StoryFn<CardProps> = (args) => (
           <p>Card content</p>
         </Col>
         <Col width="auto">
-          <Separator axis="vertical" color="accent" variant="dotted-small" fullWidth />
+          <Separator axis="vertical" color="accent" variant="dotted-small" isStretched />
         </Col>
         <Col>
           <p>Card content</p>