+++ libs/react-components/src/tedi/index.ts
@@ -6,11 +6,14 @@ export * from './components/heading-with-icon/heading-with-icon';
 export * from './components/vertical-spacing';
 export * from './components/list';
 export * from './components/grid';
+export * from './components/skeleton';
 export * from './components/tag/tag';
 export * from './components/label/label';
 export * from './components/status-badge/status-badge';
+export * from './components/separator/separator';
 export * from './components/buttons/closing-button/closing-button';
 export * from './components/print/print';
+export * from './components/buttons/button/button';
 
 export * from './helpers';