+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [9.0.0-rc.7](https://github.com/TEHIK-EE/tedi-design-system/compare/v9.0.0-rc.6...v9.0.0-rc.7) (2024-11-11)
+
+
+### Bug Fixes
+
+* **icon:** Material symbols css overriding icon--block css [#276](https://github.com/TEHIK-EE/tedi-design-system/issues/276) ([#277](https://github.com/TEHIK-EE/tedi-design-system/issues/277)) ([65f5741](https://github.com/TEHIK-EE/tedi-design-system/commit/65f5741723d10e33e23f46cd405d3b839ca09f85))
+
+# [9.0.0-rc.6](https://github.com/TEHIK-EE/tedi-design-system/compare/v9.0.0-rc.5...v9.0.0-rc.6) (2024-11-07)
+
+
+### Features
+
+* **separator:** New TEDI-Ready component Separator [#206](https://github.com/TEHIK-EE/tedi-design-system/issues/206) ([#258](https://github.com/TEHIK-EE/tedi-design-system/issues/258)) ([cd57691](https://github.com/TEHIK-EE/tedi-design-system/commit/cd576912434f2599c93120686f96d9a0173b0cc2))
+
 # [9.0.0-rc.5](https://github.com/TEHIK-EE/tedi-design-system/compare/v9.0.0-rc.4...v9.0.0-rc.5) (2024-11-05)